restored net code to a working state. Remove all Linux stuff that didn´t work and caused issues, and restored newlib / DSWIFI net code. Every net related functionality works as intended.

+++ newlib-nds/newlib-nds-headers/sys/types.h
@@ -152,7 +152,7 @@ typedef	char *	caddr_t;
     defined(__sparc__) || defined(__SPU__)
 typedef	unsigned long	ino_t;
 #else
-typedef	unsigned int	ino_t;
+typedef	unsigned short	ino_t;
 #endif
 #endif /*__CYGWIN__*/
